cleans up logger class
